feat(calendar): add SelectedDateEvents component for displaying events on selected date
- Introduced SelectedDateEvents component to encapsulate event display logic.
- Updated CalendarPage to utilize SelectedDateEvents for better separation of concerns.
- Enhanced event handling in handleAddEvent to trim titles and set dates correctly.

src/app/dashboard/calendar/page.jsx
src/components/dashboard/calendar/SelectedDateEvents.jsx
